Moving `KeycloakConfig` and `SecurityConfig` to keycloak package
